Configuration options for the RC session refresh
Add a configuration option for the periodic java-script session refresh
triggered by a setInterval() timer. Rationale: Roundcube (the email app,
not the OC app) ignores any session refresh during the first half of its
session life-time. Hence the refresh-interval has to be set somewhat
smaller then the configured session life-time of the RCmail instance in
use (default is 10 minutes, if I remember correctly). The "somewhat" has
to account for network latencies which are inevitabl in a "real"
environment. The other approach would be to refresh the session really
frequently, but that causes some network traffic. The login process is
also not exactly what I would call a "cheap" operation.

Add also a check-box in order to disable the cron-based refresh method.
I think in the default OC configuration ("AJAX" "cron" method) the cron
refresh is somewhat useless, as it is only triggered when the user
actually interacts with the browser. With the other two methods (web
"cron" or a real cron) one also would have to configure the session
life-time of the Roundcube-mail instance in use to be larger than 30
minutes, as the cron job runs only every 15 minutes.