Fix app-setting for rcHost
4627c1d5d56e6f5834e6cd73c3aee7df4a88e144 fixed the bug that the host
setting could not be reset to empty if it was once set. However, parts
of the other relied on the setting for rcHost not to be empty, but to be
non-existing.

In principle it would be able to remove the rcHost config key, however,
that functionality is somewhat hidden in the OC-lib, therefore change
the RC code to check for rcHost app-setting being empty, and then use
OC::serverHost().